refactor(forgejo,readeck): use fetch_and_deploy_codeberg_release function (#11460)

+++ ct/forgejo.sh
@@ -27,35 +27,28 @@ function update_script() {
     msg_error "No ${APP} Installation Found!"
     exit
   fi
-  msg_info "Stopping Service"
-  systemctl stop forgejo
-  msg_ok "Stopped Service"
+  if check_for_codeberg_release "forgejo" "forgejo/forgejo"; then
+    msg_info "Stopping Service"
+    systemctl stop forgejo
+    msg_ok "Stopped Service"
 
-  msg_info "Updating ${APP}"
-  RELEASE=$(curl -fsSL https://codeberg.org/api/v1/repos/forgejo/forgejo/releases/latest | grep -oP '"tag_name":\s*"\K[^"]+' | sed 's/^v//')
-  curl -fsSL "https://codeberg.org/forgejo/forgejo/releases/download/v${RELEASE}/forgejo-${RELEASE}-linux-amd64" -o "forgejo-$RELEASE-linux-amd64"
-  rm -rf /opt/forgejo/*
-  cp -r forgejo-$RELEASE-linux-amd64 /opt/forgejo/forgejo-$RELEASE-linux-amd64
-  chmod +x /opt/forgejo/forgejo-$RELEASE-linux-amd64
-  ln -sf /opt/forgejo/forgejo-$RELEASE-linux-amd64 /usr/local/bin/forgejo
-  msg_ok "Updated ${APP}"
+    fetch_and_deploy_codeberg_release "forgejo" "forgejo/forgejo" "singlefile" "latest" "/opt/forgejo" "forgejo-*-linux-amd64"
+    ln -sf /opt/forgejo/forgejo /usr/local/bin/forgejo
 
-  msg_info "Cleaning"
-  rm -rf forgejo-$RELEASE-linux-amd64
-  msg_ok "Cleaned"
+    if grep -q "GITEA_WORK_DIR" /etc/systemd/system/forgejo.service; then
+      msg_info "Updating Service File"
+      sed -i "s/GITEA_WORK_DIR/FORGEJO_WORK_DIR/g" /etc/systemd/system/forgejo.service
+      systemctl daemon-reload
+      msg_ok "Updated Service File"
+    fi
 
-  # Fix env var from older version of community script
-  if grep -q "GITEA_WORK_DIR" /etc/systemd/system/forgejo.service; then
-    msg_info "Updating Service File"
-    sed -i "s/GITEA_WORK_DIR/FORGEJO_WORK_DIR/g" /etc/systemd/system/forgejo.service
-    systemctl daemon-reload
-    msg_ok "Updated Service File"
+    msg_info "Starting Service"
+    systemctl start forgejo
+    msg_ok "Started Service"
+    msg_ok "Updated successfully!"
+  else
+    msg_ok "No update required. ${APP} is already at the latest version."
   fi
-
-  msg_info "Starting Service"
-  systemctl start forgejo
-  msg_ok "Started Service"
-  msg_ok "Updated successfully!"
   exit
 }
